requirements.txt
arabic-reshaper==3.0.0
asgiref==3.9.1
asn1crypto==1.5.1
attrs==25.4.0
autobahn==24.4.2
Automat==25.4.16
babel==2.17.0
Brotli==1.1.0
certifi==2025.10.5
cffi==2.0.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.4.3
constantly==23.10.4
coverage==7.10.7
crispy-bootstrap5==2025.6
cryptography==46.0.2
cssselect2==0.8.0
daphne==4.0.0
Django==5.2.6
django-appconf==1.1.0
django-crispy-forms==2.4
django-extensions==4.1
django-money==3.5.4
django-phonenumber-field==8.1.0
django-select2==8.4.1
django-simple-history==3.10.1
djangorestframework==3.16.1
dotenv==0.9.9
environs==3.0.0
Faker==37.11.0
fonttools==4.60.1
freetype-py==2.5.1
gunicorn==23.0.0
html5lib==1.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
lxml==6.0.2
marshmallow==4.0.1
meson==1.9.1
meson-python==0.18.0
msgpack==1.1.2
mysqlclient==2.2.7
oscrypto==1.3.0
packaging==25.0
phonenumberslite==9.0.14
pillow==11.3.0
psycopg2-binary==2.9.10
py-moneyed==3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.28.0
pycparser==2.23
pydyf==0.11.0
pyHanko==0.31.0
pyhanko-certvalidator==0.29.0
pyOpenSSL==25.3.0
pypdf==6.1.1
pyphen==0.17.2
pyproject-metadata==0.9.1
python-bidi==0.6.6
python-decouple==3.8
python-dotenv==1.1.1
PyYAML==6.0.3
read-env==1.1.0
redis==6.4.0
reportlab==4.4.4
requests==2.32.5
rlPyCairo==0.4.0
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
svglib==1.6.0
tinycss2==1.4.0
tinyhtml5==2.0.0
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
weasyprint==66.0
webencodings==0.5.1
wheel==0.45.1
whitenoise==6.11.0
xhtml2pdf==0.2.17
zope.interface==8.0.1
zopfli==0.2.3.post1
